Add note on Linux driver that I should have added a while ago
A few days after I published this repo, I went looking through the Linux
kernel to see if there were any TV tuner dongles supported in the kernel
with a protocol similar to the one this dongle uses. Lo and behold,
there was, and if I had recognized that from the beginning I probably
could have saved myself hours of work. That said, even though that
driver exists, it will still be a good amount of work to add the
necessary tuner and demodulator drivers.